backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0
openai==1.35.3
python-dotenv==1.0.1
httpx==0.27.0
langchain-text-splitters==0.2.2
numpy==1.26.4
email-validator==2.1.1
PyMuPDF==1.24.5
